artcommon/requirements.txt
google-api-core
google-auth
google-cloud-bigquery
google-cloud-core
google-crc32c
google-resumable-media
googleapis-common-protos
opentelemetry-api
